feat(ai): add vector storage capabilities with PostgreSQL support
This commit adds vector storage functionality to the ai package with the following changes:

- Add vector storage interfaces and types for managing vector embeddings
- Implement PostgreSQL provider using pgvector extension
- Add vector validation utilities for data validation
- Change embedding type from float64 to float32 for compatibility
- Update dependencies to support vector storage features

The new vector storage system provides:
- Collection management (create, delete, list)
- Document operations (upsert, get, delete)
- Vector similarity search with configurable options
- Support for metadata and custom fields
- Validation for collections and documents
- Error handling with detailed error types

Dependencies added:
- github.com/google/uuid v1.6.0
- github.com/lib/pq v1.10.9
- github.com/pgvector/pgvector-go v0.2.2

+++ ai/embedding_test.go
@@ -26,7 +26,7 @@ func TestEmbeddingService_GenerateEmbedding(t *testing.T) {
 				Data: []EmbeddingObject{
 					{
 						Object:    "embedding",
-						Embedding: []float64{0.1, 0.2, 0.3},
+						Embedding: []float32{0.1, 0.2, 0.3},
 						Index:     0,
 					},
 				},
@@ -46,12 +46,12 @@ func TestEmbeddingService_GenerateEmbedding(t *testing.T) {
 				Data: []EmbeddingObject{
 					{
 						Object:    "embedding",
-						Embedding: []float64{0.1, 0.2, 0.3},
+						Embedding: []float32{0.1, 0.2, 0.3},
 						Index:     0,
 					},
 					{
 						Object:    "embedding",
-						Embedding: []float64{0.4, 0.5, 0.6},
+						Embedding: []float32{0.4, 0.5, 0.6},
 						Index:     1,
 					},
 				},
